fix app and rename file back to router
