+++ frontend/src/pages/emailVerification/Status.tsx
@@ -8,7 +8,7 @@ import {
 import { Image, ImageProps } from 'codeforlife/lib/esm/components';
 import { ThemedBox, ThemedBoxProps } from 'codeforlife/lib/esm/theme';
 
-import { paths } from '../../app/routes';
+import { paths } from '../../app/router';
 import { themeOptions } from '../../app/theme';
 
 const Status: React.FC<{
