+++ frontend/src/pages/error/ErrorProps.tsx
@@ -5,7 +5,7 @@ import {
 
 import { ImageProps } from 'codeforlife/lib/esm/components';
 
-import { paths } from '../../app/routes';
+import { paths } from '../../app/router';
 import KirstyImage from '../../images/kirsty.png';
 import NigelImage from '../../images/nigel.png';
 import DeeImage from '../../images/dee.png';
