+++ frontend/src/pages/home/CodingClubs.tsx
@@ -8,7 +8,7 @@ import {
   ChevronRight as ChevronRightIcon
 } from '@mui/icons-material';
 
-import { paths } from '../../app/routes';
+import { paths } from '../../app/router';
 
 const CodingClubs: React.FC = () => {
   return (
