+++ frontend/src/pages/homeLearning/HomeLearning.tsx
@@ -9,7 +9,7 @@ import {
 
 import Page from 'codeforlife/lib/esm/components/page';
 
-import { paths } from '../../app/routes';
+import { paths } from '../../app/router';
 import AboutRR from './AboutRR';
 import Levels from './Levels';
