+++ frontend/src/pages/privacyNotice/ForChildren/ForChildren.tsx
@@ -6,7 +6,7 @@ import {
 
 import Page from 'codeforlife/lib/esm/components/page';
 
-import { paths } from '../../../app/routes';
+import { paths } from '../../../app/router';
 import TableOfContents from '../../../components/TableOfContents';
 import WhoWeAre from './WhoWeAre';
 import PersonalInfo from './PersonalInfo';
