+++ frontend/src/pages/teacher/RapidRouter.tsx
@@ -8,7 +8,7 @@ import {
   ChevronRight as ChevronRightIcon,
   OpenInNew as OpenInNewIcon
 } from '@mui/icons-material';
-import { paths } from '../../app/routes';
+import { paths } from '../../app/router';
 import TeacherSection from './TeacherSection';
 
 const RapidRouter: React.FC = () => {
