+++ frontend/src/pages/teacher/Resources.tsx
@@ -7,7 +7,7 @@ import {
   ChevronRight as ChevronRightIcon
 } from '@mui/icons-material';
 
-import { paths } from '../../app/routes';
+import { paths } from '../../app/router';
 import TeacherSection from './TeacherSection';
 
 const Resources: React.FC = () => {
